+++ test/featherlight_gallery_test.js
@@ -154,8 +154,8 @@ var expect = chai.expect;
     describe('.current', function() {
       it ('only returns actual featherlight gallery instances', function() {
         $.featherlight('<p>This is a test<p>');
-        expect($.featherlight.current()).to.not.be.undefined;
-        expect($.featherlightGallery.current()).to.be.undefined;
+        expect($.featherlight.current()).to.not.be.null;
+        expect($.featherlightGallery.current()).to.be.null;
       });
     });
   });